refactor: 전체 코드 컴포넌트 분리
