+++ src/main.tsx
@@ -5,6 +5,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 
 import App from './App.tsx';
+import { EventFormProvider } from './context/FormContext.tsx';
 
 const theme = createTheme();
 
@@ -12,9 +13,11 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <ThemeProvider theme={theme}>
       <CssBaseline />
-      <SnackbarProvider>
-        <App />
-      </SnackbarProvider>
+      <EventFormProvider>
+        <SnackbarProvider>
+          <App />
+        </SnackbarProvider>
+      </EventFormProvider>
     </ThemeProvider>
   </React.StrictMode>
 );
